pyproject.toml
[project]
name = "restaurant-agent"
version = "0.1.0"
description = "AI restaurant recommendation agent with fraud and viral-risk detection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.1",
    "requests==2.32.3",
    "python-dotenv==1.0.1",
    "pydantic==2.12.0",
    "pytest==8.3.5",
    "httpx==0.28.1"
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
